Actually, now that I think about it, this is probably wrong.  I'm reverting this change for now.
What I observed is that in a ZODB based app, the ZODB connection was closed
when I let the closer fall out of scope.  I see from examining the code, 
though, that this is a side effect of the environ that the closer hangs on
to getting deleted.  There is nothing here that pops the registry when the
closer is deleted.  You still have to call the closer for that to happen.