Possible fix for view_config() referencing old registry
view_config decorators are called only once for an entire test suite execution.
After tearDown() is called once, further calls to config.scan() fail to update
the registry of the associated renderer of a view, which causes test cases to
fail.

This patch contains a test case to highlight the problem, and a proposed fix.