Use simplejson when available.
Both modules are actually the same, one integrated in the Python standard
library and the other distributed independently.  If somebody wants to take
advantage of the new JSON serialization speed improvements without having to
wait for the next Python release, it is as easy as installing the simplejson
module.

The API is exactly the same so it is easy to exchange one for the other.