Fixes issue #637.
This fix adds a `depth` argument to the `@view_config` decorator
to make use of this feature in Venusian back-end.  This enables
use of the `@view_config` decorator in local functions.

Note that this is a big work around for the fact that the Venusian
implentation tries to access the call stack (at decorator callback
invocation time) in order to check if the decorator is used in a
class or module context.