preliminary fix for refactorings from 4d2602ce
with the change from `__text__` to `text()` the code putting together the output used when `debug_routematch` is set was broken.  it still uses `.__text__`.  the same is true for the `pviews` script.  however, simply changing things to use `.text()` breaks `test_views_command_single_view_route_with_route_predicates` and i'm not sure if the function used for a predicate there can be changed into a class, or if function predicates should still be supported...