update pviews to work with the bootstrapped request
had to fix DummyBootstrap a bit because it fails pretty hard at handling
part of bootstrap contract in which the request passed in should be
enhanced with the registry, and returned. In some cases the wrong
request or the wrong registry were being returned if the test case
actually specified a registry or request.