notfound and forbidden decorators were ignoring view_defaults
This could be fixed in other ways but the basic problem is that because
config.add_notfound_view and config.add_forbidden_view have actual
signatures instead of *args, **kwargs, the arguments are squashing the
view_defaults which are applied later on the call to config.add_view.

Basically, by the time the args get to config.add_view, they look
explicit when they are not.

fix #1173