Proxy attribute access to the underlying object
This is a smart proxy, in that if the attribute being accessed is a
function, it will return a wrapper so that after the attribute has been
called on the original value we make sure to return ourselves.

This way not_('TEST').lower() will return an object that is the same as
if the user has typed not_('test'). This is mainly of use in predicates
that get lowercased automatically, for example in @view_config, the
AcceptPredicate is lower()'ed.

Closes #1391