Add an implicit_content_type attribute to Response
Use the fact that Response's content_type is a property, and modify the
set function so that it flips implicit_content_type from True to False.

This works because WebOb's __init__ sets the content-type directly on
the headerlist, rather than using the getter/setter on the property.
Since the only time that the setter would get called is due to user
code, we can flip a flag.

This is so going to break if someone modifies WebOb...