Remove __providesBy__ attribute from being copied
This change reverts part of what
https://github.com/Pylons/pyramid/issues/1212 changed. We attempt to
copy information used by zope.interfaces to the new class to avoid
creating new objects that could possibly leak.

Unfortunately copying over __providedBy__ causes issues with the
sub-class after modification no longer implementing interfaces, for more
information see:
https://github.com/Pylons/pyramid/issues/1529#issuecomment-73373581

Once __providedBy__ is removed from the list, this is no longer the case
and the new class that is created in `_set_properties` properly report
that it implements the interfaces it used to implement before.