Add AlchemyDox Template
Add an additional sqlalchemy template that contains docstrings as
out-of-the-box documentation within a new alchemy template without
modifying the original alchemy template. This entry_point, alchemydox,
has been added to setup.py and is accessible for creating a new
application based on this template. Though, not necessary, the
application documentation, if created with Doxygen, is best created with
the doxypy.py script as an INPUT_FILTER. Inputted documentation is still
useful even if not using Doxygen.