Fix Pyramid against WebOb >=1.5.0
Due to changes in WebOb we may no longer create a Response() object with
an invalid status (i.e. 'None None'). For the top-level HTTPException if
it is not correctly overriden in child classes then we now use '520
Unknown Error'.

While not an official RFC error, this has been used by Microsoft's Azure
as well as CloudFlare to signify an error that doesn't have a more
appropriate error message.