Rename predicates to view_options
We also use the previously defined names() method on PredicateList to
fetch the list of predicates and pull out all predicate values into it's
own dictionary.

Then the rest of the values are the view options for the view
derivations. At the moment we do not check if the options matches a
valid view derivation. No seatbelts for view options.