View derivations get their name as a kwarg as a value
Instead of passing in a default to the view derivation, we attempt to
pull it out of the options kwarg, if that doesn't exist, then we pass in
the default.

This should make it simpler to program view derivations because now you
don't need to check to see if the kwarg is different from the default.