rendered_view is now part of the pipeline
This means that it can be overriden, and is no longer static. With great
power comes great responsibility...

This also allows the user to insert view derivations over the
rendered_view. This would allow the same sort of functionality as
the BeforeRenderer event, by allowing the derivation to modify the
Response/dictionary that the view returned.