Add 'pyramid.session.BlowfishPickleSerializer'.
This class is an encrypting Webob cookie serializer, which uses the pickle
protocol to dump Python data to bytes, then encrypts the value symmetrically
using the Blowfish algorithm.

To encrypt sessions, pass an instance of this class as the 'serializer'
argument to 'pyramid.session.BaseCookieSessionFactory'.

Before using this serializer, install 'pyramid[encrypted_cookies]' to
get the necessary dependency ('pycrypto').