Do not use a set for accept dedup in MultiView.add
Starting python3.3, string hash are randomised by default. Meaning the
self.accept list order was random.

In case of a MultiView accepting multiple types, the first matching
type of that list is used.

The result was that the method actually used was random based on the
initial seed.

Example bellow. When using GET without an accept header (or */*), the
method called was randomly get_json or get_text

```
@view_defaults(route_name='example')
class ExampleView(object):
    def __init__(self, request):
        self.request = request

    @view_config(renderer='json', accept='application/json')
    def get_json(self):
        print('GET JSON')
        return 'json'

    @view_config(renderer='string', accept='text/plain')
    def get_text(self):
        print('GET TEXT')
        return 'text'
```