convert response callbacks into a tween
There is a backward incompatibility here in how tweens are defined. This
is consistent with the new view deriver API but may present some
frustration for anyone currently defining a tween **under** the EXCVIEW
as all tweens now default to **over** the EXCVIEW.

This incompatibility is not strictly required to implement this feature,
but if we don't then:

1) Tween ordering is almost completely dependent on the user's
   environment and sorting behavior. If they add another tween then the
   addon tween (if left unspecified) may be shifted to a bad place.

2) Another incompatibility may surface in which a user's tween would
   show up OVER the response callbacks tween, which would lead to
   inconsistent behavior versus the previous pipeline where callbacks
   were always executed after tweens.

This patch shifts tweens to be more fully specified which should be
generally considered a good thing.

Side note, the response callbacks and excview tweens are defined
with fallbacks incase one of them is overridden or unavailable for some
reason by using a list of dependencies. We should apply this pattern to
the view derivers as well to make them easier to override.