Avoid setting Content-Encoding header for static view responses.
This was causing clients to decode the content of gzipped files
when downloading them.

We discussed this on IRC and @mmerickel pointed out that it may be a goal in some cases
to serve gzipped precompiled assets (i.e. CSS/Javascript) with this header.
But that's a new feature that would require thought about how to
specify which files to serve that way. And it can already be
implemented for a project using a tween. This just aims to
fix the existing use case of serving files for download.