Add a warning: view derivers only take effect for view configured afterwards
I ran into this because I had assumed Pyramid maintains a chain of view derivers in its config that is applied whenever a view is called. Actually, the chain is made and configured to be used by a view at the time that this view is added to the config - with whatever view derivers are known at that time.