Update debugtoolbar.rst
if a tutorial follower decide to get rid of response ( html ) and just write a simple string like ( it worked ) , the toolbar wont show unless an error is introduced. I would like to contribute but I don't know where that fits since it is mentioned later in the analysis but it might confuse new comers.